Refactor the simulation pipeline
Deprecated the bloated pyzx.simulate.py file in favour of a structured pyzx.simulation folder with decompositions and decomp strategies separated into individual files, etc.
