+++ pyzx/__init__.py
@@ -47,6 +47,7 @@ from . import editor
 from . import routing
 from . import local_search
 from .routing.parity_maps import CNOT_tracker
+from . import simulation
 
 # some common scalars
 from .graph.base import Scalar
